requirements.txt
aiosqlite==0.22.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.124.4
google==3.0.0
google-auth==2.45.0
google-genai==1.56.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iso8601==2.1.0
mutagen==1.47.0
numpy==2.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
PyJWT==2.10.1
pypika-tortoise==0.6.3
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
requests==2.32.5
rsa==4.9.1
scipy==1.16.3
sniffio==1.3.1
soupsieve==2.8.1
starlette==0.50.0
tenacity==9.1.2
tortoise-orm==0.25.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
websockets==15.0.1
